Move wrappers to a separate file
